Fixes Execution Directory reference and moved utils to unskript_utils.py
